requirements.txt
# Core Data Engineering Dependencies
pyspark==3.5.0
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23

# Streaming & Messaging
kafka-python==2.0.2
confluent-kafka==2.3.0

# Cloud & Storage
boto3==1.34.0
aws-glue-libs==4.0.0
pyspark[sql]==3.5.0

# Orchestration
apache-airflow==2.7.3
celery==5.3.4

# ML & AI
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
transformers==4.35.2
langchain==0.0.350
openai==1.3.7

# Data Processing
apache-airflow-providers-amazon==8.7.0
apache-airflow-providers-postgres==5.6.0

# Monitoring & Visualization
plotly==5.17.0
dash==2.14.2
streamlit==1.28.2

# Database & Querying
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# Testing & Quality
pytest==7.4.3
pytest-cov==4.1.0
great-expectations==0.18.5

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
schedule==1.2.0

# Live Streaming Dependencies
websockets==12.0
asyncio
